requirements.txt
apache-airflow-providers-snowflake==3.0.0